pyproject.toml
[project]
name = "condor"
version = "0.1.0"
description = "Telegram bot for Hummingbot trading via Backend API"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[job-queue]",
    "hummingbot-api-client==1.4.1",
    "python-dotenv",
    "PyYAML",
    "aiohttp",
    "watchfiles",
    "kaleido",
    "plotly",
    "pandas",
    "geckoterminal-py",
    "mcp",
    "fastapi",
    "uvicorn[standard]",
    "python-jose[cryptography]",
    "faster-whisper",
    "matplotlib>=3.10.8",
    "scipy>=1.17.1",
    "pydantic-ai[mcp]",
    "beautifulsoup4>=4.14.3",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "black", "isort", "pre-commit"]

[tool.isort]
profile = "black"
